backend/requirements.txt

Django==4.2.4
djangorestframework==3.14.0
django-filter==23.1
django-celery-beat==2.5.0
django-redis==5.4.0
celery==5.3.1
redis==5.0.1
psycopg2-binary==2.9.7
#dr-f==0.0.1 # Placeholder for Django Rest Framework

djangorestframework-simplejwt==5.3.0



django-cors-headers==4.2.0



python-json-logger==2.0.7


